refactor: move vela-api tests from src to test directory
Move test files from src/ to test/ directory to avoid potential issues:
- test/dynamodb.test.ts (from src/)
- test/routes/chat-history.test.ts (from src/routes/)
- test/routes/llm-chat.test.ts (from src/routes/)

This keeps tests separate from production code while maintaining
the same directory structure for organization.

All 21 tests passing.
